Add game and https certification to deployment scripts
I created the dist folder with `npm run build`, which I then put on
the server home directory. A symbolic link was then made connecting
/var/www/html/game to ~/dist.

package-lock.json was automatically generated, but the online doc
says it should be put in repository. Not sure what it does, something
about common dependencies?